Create Keras_NN_text_gen_trainer.py
A fairly simple RNN trainer for the Keras_NN_text_generator.py. It is character based, generating text one character at a time rather than one word at a time. It is small enough to train on a Mac or PC, but just barely (it took 30 hours on mine, after which I decided to run it on my university's High Performance Computing Cluster).